// src/pages/dashboard/components/attendanceLog/data.d.ts
export type AttendanceLogItem = {
  key: number;
  type: 'Schedule' | 'Hours' | 'Time';
  [key: string]: any;
};

export type Pagination = {
  total: number;
  pageSize: number;
  current: number;
};

export type Params = {
  status?: string;
  name?: string;
  desc?: string;
  key?: number;
  pageSize?: number;
  currentPage?: number;
  filter?: Record<string, any[]>;
  sorter?: Record<string, any>;
};